Fixing some circular imports, and pushing for review.
* Changing URL in read me for discord.com instead of discordapp.com
