+++ README.md
@@ -2,7 +2,7 @@
 
 [![PyPI](https://img.shields.io/pypi/l/disco-py.svg)](https://pypi.python.org/pypi/disco-py/)
 
-Disco is an extensive and extendable Python 3.x library for the [Discord API](https://discordapp.com/developers/docs/intro). Disco boasts the following major features:
+Disco is an extensive and extendable Python 3.x library for the [Discord API](https://discord.com/developers/docs/intro). Disco boasts the following major features:
 
 - Expressive, functional interface that gets out of the way
 - Built for high-performance and efficiency
