+++ disco/types/voice.py
@@ -1,4 +1,3 @@
-from disco.types.guild import GuildMember
 from disco.types.base import SlottedModel, text, Field, snowflake, cached_property, datetime
 
 
@@ -6,7 +5,6 @@ class VoiceState(SlottedModel):
     guild_id = Field(snowflake)
     channel_id = Field(snowflake)
     user_id = Field(snowflake)
-    member = Field(GuildMember)
     session_id = Field(str)
     deaf = Field(bool)
     mute = Field(bool)
@@ -28,6 +26,10 @@ class VoiceState(SlottedModel):
     @cached_property
     def user(self):
         return self.client.state.users.get(self.user_id)
+
+    @property
+    def member(self):
+        return self.guild.get_member(self.user_id)
 
 
 class VoiceRegion(SlottedModel):
